correcting CE and add ability to have float numbers

+++ src/myGui/Divider.java
@@ -3,12 +3,12 @@ package myGui;
 import java.util.Stack;
 import java.util.EmptyStackException;
 public class Divider extends Operator {
-	public Divider(ResultPresenter rp,Stack<Integer> st) {
+	public Divider(ResultPresenter rp,Stack<Float> st) {
 		super(rp,st);
 	}
 	public void operate(){
 		try {
-			Integer n = st.pop();
+			Float n = st.pop();
 			if (n!=0 && st.size() > 0) {st.push( st.pop()/n);}
 			else if (n == 0){rp.error("error dia 0,'to  0 diagrafike apo thn stoiba ksnaprospathise"); }//
 			else {
